requirements.txt
Flask
gspread
oauth2client
gunicorn
python-dotenv